pyproject.toml
[project]
name = "dockerproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "psycopg2-binary>=2.9.10",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.35.0",
]
